Convert Skeleton to TS.
