--- packages/components/src/Skeleton/index.js
@@ -1,2 +1,0 @@
-export { default } from './Skeleton';
-export { default as Skeleton, SkeletonSize } from './Skeleton';
